requirements.txt
amqp==2.6.0
asgiref==3.2.10
beautifulsoup4==4.9.1
billiard==3.6.3.0
celery==4.4.6
certifi==2020.6.20
chardet==3.0.4
coverage==5.2.1
DAWG-Python==0.7.2
dj-database-url==0.5.0
Django==3.0.8
django-crispy-forms==1.9.1
django-filter==2.4.0
django-redis==4.12.1
docopt==0.6.2
dostoevsky==0.6.0
fasttext==0.9.2
flashtext==2.7
flower==0.9.5
future==0.18.2
gunicorn==20.0.4
humanize==2.5.0
idna==2.10
importlib-metadata==1.7.0
kombu==4.6.11
numpy==1.19.3
prometheus-client==0.8.0
psycopg2-binary==2.8.6
pybind11==2.6.2
pymorphy2==0.8
pymorphy2-dicts==2.4.393442.3710985
python-decouple==3.3
pytz==2020.1
razdel==0.5.0
redis==3.5.3
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
tornado==6.0.4
urllib3==1.25.9
vine==1.3.0
vk==2.0.2
whitenoise==5.2.0
zipp==3.1.0
